Initial commit with systemd support
NOT WORKING ON SYSTEMD 253 AND ABOVE.
NOT VERIFIED TO BE COMPATIBLE WITH EXISTING CRYPTTABS.

In its current state, this crashes systemd-253 and above with the error message "Failed to start up manager" during early boot, which halts startup.

Support for systemd also requires the pgp keyfile to be specified as `pgp-keyfile=...` as an option in the crypttab entry, instead of in the keyfile column. (Otherwise, systemd-decrypt will try to load it as an ordinary keyfile.)

Documentation and migration tools will be updated once this gets closer to release quality.